% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdda.R
\name{rex_extract}
\alias{rex_extract}
\title{Automatic discovery of regular-expression constraints}
\usage{
rex_extract(strings)
}
\arguments{
\item{strings}{a character vector of value for which a matching regular
expression is wanted}
}
\description{
Automatic discovery of regular-expression constraints
}
\examples{
  rex_extract(c("EH1", "7JQ", "WC1", "4AA", "G2", "3PQ"))
  rex_extract(c("123-AA-971", "12-DQ-802", "198-AA-045", "1-BA-834"))
}

